requirements.txt
albucore==0.0.15
albumentations==1.4.14
annotated-types==0.7.0
autocommand==2.2.2
backports.tarfile==1.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
eval_type_backport==0.2.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
huggingface-hub==0.24.6
idna==3.7
imageio==2.35.1
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.9.2
more-itertools==10.3.0
networkx==3.3
numpy==1.26.4
opencv-python-headless
ordered-set==4.1.0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
progress==1.6
pydantic==2.9.1
pydantic_core==2.23.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.4
scikit-image==0.24.0
scipy==1.14.0
six==1.16.0
spikingjelly==0.0.0.0.14
syops==0.0.5
tifffile==2024.8.30
timm==1.0.8
tomli==2.0.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchsummary==1.5.1
torchvision==0.14.1+cu117
tqdm==4.66.5
typeguard==4.3.0
typing_extensions==4.12.2
urllib3==2.2.2
zipp==3.19.2
